Iterate Cluster Minion configs
Various improvements intended to enhance modularity of Cluster Minions
and prepare the integration of VxGW Service. The main chantes are trying
to keep configuration for different data sources separate and
independent from specific services, as well as providing independent
Guice bootstrapping for them.

Adding also instructions on how to run the cluster in
brain/midonet-brain/README.

Change-Id: Iaf6fbafa20407d751ebe96a2445e254979bbe5be
Signed-off-by: Galo Navarro <galo@midokura.com>