C3PODaemonTest starts the service exactly once.
BeforeAndAfter trait runs the before{} clause before *each* test, but
we need to start the daemon just once before all the tests. For that
we need to use BeforeAndAfterAll trait and override before/afterAll().

Also, C3PODaemonTest now deletes all the task rows from the task DB
and then sends a flush task before each test run so that the task DB
and the Storage would be properly initialized.

Change-Id: I3921caf9aeb55341d394ee60a812760e66b97ff1