Update patch set 3
Patch Set 3:

This file is used to locally persist the UUID of a VXGW instance, such that it can reclaim previously owned VTEPs after a restart. Currently, there is no other information written to it. Although the user can modify this file to change the UUID of the VXGW service, the recommended approach is to modify the "host_uuid" property in the service's configuration file.

Patch-set: 3