Customizing timeouts with eventually blocks
This patch proposes to use the MidonetEventually trait to customizes the
"PatienceConfig" used in "eventually" blocks. A patience
configuration contains a timeout and an interval parameter. Timeout
defines how long the "eventually" block can execute before a timeout
occurs. The interval specifies how much time should pass between each
trial.

The timeout value is set to 3 seconds (the maximum timeout used
in the tests) and the interval is 15 milliseconds (the default value in
PatienceConfiguration).

This patch also updates tests where an "eventually block" is used to make
them use MidonetEventually.

Change-Id: I92534dcd49c6bdac3bba9d936273099f40ef256b
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>