Rearch: Refactor port status to cluster storage
This patch updates the StateStorage trait and it Zookeeper
implementation injected into the agent via the StateStorageModule,
which exposes a middle-layer API between the agent and the cluster
storage. This class is compatible with both the ZOOM storage and
the legacy ZK storage.

This allows the activation status of a port to be changed in
both the legacy ZK storage and the re-architected cluster storage.

This patch also modifies the notification of the port activation/
deactivation, which is now exposed as an observable, and where
an AKKA actor mixing the ReactiveActor trait can subscribe
directly, receiving the notifications in its receive() method.

Ref: MN-3126

Change-Id: I461bdb8cd09f7fe35f5dc18f5616f762debfafd8
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>