Don't create user mode flows with userspace fields
This patch stops creating userspace flows for packets that contain
userspace-only fields such as ICMP ID or data. This solves a bug where
mangling of fields on a packet encapsulated in an ICMP packet, such as
reversing a NAT, was only being performed for the first simulated
packet and not for subsequent packets that matches the wildcard flow
table.

Change-Id: Ifff2694c44abaabfcfb93a8bdd81e29591e70728
Signed-off-by: Duarte Nunes <duarte@midokura.com>