Refactor management of VxGW logical switches
This patch provides a new implementation of the VxLAN Gateway
Service that supports bindings to multiple VTEP from the same Neutron
network. Most of the refactor simplifies the older model, by treating
the control of VTEP and MidoNet sides as decoupled and self-sufficient
as possible.  Each participant in a VxLAN Gateway (that is, a binding
between a Neutron network and a bunch of port/vlan pairs on a number of
hardware VTEPs) publishes/subscribes to/from a message bus.

Change-Id: I1d9c2c9d4d7c10c3e09622997d6e35981b5cca19
Signed-off-by: Galo Navarro <galo@midokura.com>