The VTPM exposes the new Host class
Previously the VTPM would reply to host requests by sending a host of class org.midonet.midolman.topology.rcu.Host. Since the HostMapper returns a host of class org.midonet.midolman.topology.devices.Host, this patch makes the VTPM return hosts of that same class (the HostManager continues to use rcu.Host).

Additionally, this patch removes the epoch field in class ResolvedHost since it was never used. Also, the datapath field of rcu.Host was always set to "midonet" and as a consequence devices.Host does not contain such field and the datapath "midonet" is now hard-wired in several places.

Change-Id: I66021e179352cd35b0520068b113b109cf281bc6
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>