Update patch set 2
Patch Set 2: Code-Review+2

> Looks good, but is the text always Json? If it is, maybe it would
 > be better to s/txt/json/? It's just a suggestion. Feel free to
 > check in anyway if you don't think it's a good idea.

Thanks for taking a look, Brandon. Yes, the text is what you call "Json". I know that with "Json" you refer to something more general than Json. However, to me it refers to the specific format, and Protobuf's text format is slightly different from that. So I'd prefer to keep the names as is.

Patch-set: 2
Label: Code-Review=+2