ZOOM: Ownership support to in-memory storage
This patch implements the methods of the StorageWithOwnership trait
in the InMemoryStorage class, allowing to use the in-memory storage
to save ownership data. The implementation is compatible with, and
passes the same unit tests, as the ZookeeperObjectMapper.

To eliminate the duplication of code, and to ensure the same
transactional behavior with the ZookeeperObjectMapper, this patch
also consolidates the transaction manager in a single public
TransactionManager class, which is extended by both the
ZookeeperObjectMapper and the InMemoryStorage classes.

Ref: MN-3133

This patch includes the fix for the object identifier bug.

Fixes: MNA-115

Change-Id: I46b7bf05a5f0415d5bbf87ca91864e50e6c86427
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>