Update patch set 17
Patch Set 17:

(1 comment)

Thank you, Galo. Looks good to me. I agree that we will like do away with childrenLock. But with or without childrenLock, the ObservableNodeCache and ObsevablePathChildrenCache both still do NOT look to be thread-safe. But I agree that we can address that, if necessary, in a separate thread.

Patch-set: 17