Update MidolmanServices to Scala guidelines
This patch removes the parantheses for accessor methods in
MidolmanServices class, bringing them in line to Scala
guidelines, according to which methods which act as accessors
of any sort (either encapsulating a field or a logical property)
should be declared without parentheses except if they have
side effects.

The patch also cleans up other unnecessary parantheses and
unused imports.

Change-Id: I6c079c4c032108ba0ecce86c4d8fb71dead163fa
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>