Fix DHCPValueParser to utilise the appropriate option list
This patch fixes the issue of DHCPValueParser not to dispatch the
appropriate parser because of null filled arrays defined outside
of Code enum that fails to have references to Code enum values.

This patch also adds missing vendor extended options and now we
have the comprehensive list of all options for DHCPv4.

This fixes MNA-58 and MNA-59.

Change-Id: I4f938d7701a4522224b896f88270f7a809bc33cd
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>