Introduce dhcp_mtu in midolman.conf for PacketWorkflow
This patch introduces dhcp_mtu option in midolman.conf and
modify PacketWorkflow to use it if it's available. Otherwise
Midolman uses the default value, i.e., 1500, in the DHCP ack
packets.

This MTU value gets passed to DatapathController when it starts
via the config file. VMs are notified this value, from which is
subtracted the tunnel overhead if the host belongs to any tunnel
zone, through DHCP ack packets instead of the underlay interface
MTU-based value.

This fixes MN-3035.

Change-Id: I0f2f33d4003e6b64dd676cee489136ae820a463c
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>