OVSDB plugin update, fix physical locator sets
The old implementation was adding mutiple physical locator set entries
for the unknown-dst entry of each logical switch. The hardware switch
doesn't like this, it will just take one of the entries and ignore the
rest.

This patch fixes this. When adding a new ip to the unknown-dst locators,
we'll remove the mcast_macs_remote entry and the associated physical
locator set, and create new ones. We can't mutate, because OvsDB seems
quite restrictive in what columns you can mutate and cannot.

Change-Id: Icd8d98bad099cb0893f108d697a4f0fb179e76c5
Signed-off-by: Galo Navarro <galo@midokura.com>