Update patch set 6
Patch Set 6:

I don't believe this patch actually addresses the underlying problems, which are data races. Prove of that is the number of times midonet_gate still failed. Increasing the timeouts means we will have to wait more for a build to fail, because indeed they will never complete (and 15 millis is already ages in CPU time).

Patch-set: 6