Changes to the Rule ProtoBuf
This patch modifies the rule protocol buffer to make it compatible with
the Zoom storage. A rule has an id, a type, an action, a chain_id, and a
condition. The remaining fields are specific to the different rule
types, and are implemented as a 'oneof' field.

Additionally, this patch provides a way to convert a protocol
buffer to an object of class org.midonet.midolman.rules.Rule. The
position field of a rule is removed from the Rule class as it was not
used.

Change-Id: I96933701630a2e39e7e4cb02bdd713d02a7fbf1e
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>