Ensure NatTarget ports are always within range
This patch fixes a bug (and prevents future ones) whereby a class was
not properly trimming the port fields in NatTarget. More specifically,
when creating a rule with NatTargets, the Neutron plugin or the user
can write -1 to mean the maximum allowed port value. This is stored in
ZK and read from it as an integer instead of a short, breaking code
that expects the port value to be within the valid range.

Ref MN-3174

Change-Id: Ie4f4029c478ffb701759dbe6dd7751dcc7c9dfa6
Signed-off-by: Duarte Nunes <duarte@midokura.com>