Create dependency mapping of asm:asm to org.w2.asm:asm
Guice pulls in version 3.1 of asm, while jmh pulls in 5.0.3. Gradle
doesn't notice the that these are conflicting, as the group is
different. This causes generateBenchmark to fail on some systems. It has
failed on all linux hosts I've tried, but doesn't on some. It depends on
which jar gets pulled in first.

Ref MNA-3

Change-Id: Ifeb88c3dfd64383c3fbe289d98f51c523e1e87dd
Signed-off-by: Ivan Kelly <ivan@midokura.com>