Update Gradle and create dependency mapping of asm:asm to org.w2.asm:asm
Guice pulls in version 3.1 of asm, while jmh pulls in 5.0.3. Gradle
doesn't notice the that these are conflicting, as the group is
different. This causes generateBenchmark to fail on some systems. It has
failed on all linux hosts I've tried, but doesn't on some. It depends on
which jar gets pulled in first.

I've also updated the wrapper to use gradle 2.2 as the dependency
mapping functionally only exists from this release onwards.

Ref MNA-3

Change-Id: I12f4d71e1cee8abc48e8fda6abc4f8eafddc1b29
Signed-off-by: Ivan Kelly <ivan@midokura.com>