Update patch set 1
Patch Set 1: Code-Review+1

I'll +1 this but curious about the motivation behind it.  The API Guice modules are organized by features, so in case we ever want to remove a particular feature it would be easy to do so.  So serialization package was kept separate from versioning package.  By merging this, I'm guessing what you are trying to do is squash Guice modules so that we do not have to deal with dependencies across the modules, is that correct?   What was your reasoning to merge Version and Serialization modules?  Is it simply because that's where the dependency happens to exist right now in the code?  Is the plan to merge all the Guice modules together?  Or are you planning to group them in a certain way?  Anyway, feel free to merge.

Patch-set: 1
Reviewer: Ryu Ishimoto <1002740@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1