Update patch set 1
Patch Set 1:

Ryu, yes, I'm mostly trying to squash modules that generally go together (e.g, version and serialization) in order make the dep graph between modules more manageable and then refactor.

The approach I'm trying to move to is providing top-level modules such as "MidonetBackendStorage" or "LegacyBackendModule" that nest other smaller modules and move these into /cluster so that they are in a single place, offered as libraries to the other components. This is now really hard because of the vast tree of deps crossing multiple modules (midolman, -api..). 

Also, in the way, I was trying to do some rationalization of Guice usage, which IMO we abuse a lot. There are cases where we instantiate a small utility class in a single place, and instead of just instantiating it, we require binding it and the whole guice scaffolding. Another example is that by injecting absolutely everything, it's very hard to reuse a class: instantiating it requires a whole huge Guice setup so that all the tangled deps are provided. So in practise, what we're supposed to achieve with Guice (modularity), is defeated. A good proof of this are many tests, that trying to instantiate a single thing still require logs of (otherwise unnecessary) Guice bootstrapping.

Patch-set: 1