Update patch set 1
Patch Set 1:

Oh and btw, I think this change doesn't harm the simplicity to remove a feature. Removing versions in this case is simple, we can go to the SerializationModule and remove the binding. As we get more stuff unified, this should become better: bindings are provided by the cluster in one place. Now it's not the case as various modules are duplicated between -api and midolman

Patch-set: 1