Update patch set 31
Patch Set 31: Code-Review-1

Hello Alex;

The following is the contract for "update", taken from Storage.scala.

    /** 
     * Updates an object and/or its owner. For [[OwnershipType.Exclusive]], the 
     * update succeeds only if the specified owner is the current owner of the 
     * object, or if the object has been orphaned and it has no owner. The 
     * method rewrites the current ownership node in storage, if the update
     * has been successful.
     */  
    @throws[NotFoundException]
    @throws[ReferenceConflictException]
    @throws[OwnershipConflictException]
    def update[T <: Obj](obj: T, owner: ObjId,
                         validator: UpdateValidator[T]): Unit

Unfortunately it doesn't say much about when the ownership type is shared ownership. However, I'd expect that update would also succeed with the shared ownership type if you're already an owner, and I'd expect similarly with "updateOwner". You gave me an example of the shared Port ownership by Agents, where the Agent owning the Port means that the Port is active with the Agent's host. I don't know whether an Agent would ever need to update the Port status by herself, but if she needs to do so, shouldn't she be allowed if she's an owner? I don't really see the strong need for disallowing multiple ownership requests by the same owner on the same object. In that sense, as you suggested, getting rid of rewriteOwner makes sense.

Another source of confusion is that apparently update / updateOwner are also being used for requesting for the shared ownership. I think it's better to have something like getWithOwnership or addOwner for requesting ownership, which should succeed if the requested object is of the shared ownership type (and may / may not fail if the caller already has the ownership, which is a design choice). Update and UpdateOwner should always succeed whether the caller already owns it or not, and when it doesn't, we may add the caller as a new shared owner. 

What do you think?

Patch-set: 31
Label: Code-Review=-1