Move Neutron cluster APIs to REST API
The ZookeeperLockFactory is introducing a dependency on Curator into the
DataClientModule, which we want to avoid to keep the new storage
independent from the old one (and moved into /cluster).  The
ZookeeperLockFactory itself is actually only needed in the API, through
the Neutron cluster API which resides in Midolman.  The ideal would be
to move this cluster API to /cluster, where it belongs.  However this
would require also dragging the DataClient and Zk managers along, which
is a larger problem.  So we are doing here the opposite move, push the
Cluster API to the MidoNet API, taking advantage of the fact that both
Cluster API and REST API will disappear with the new storage stack.
This removes the dependency from Midolman on a plain Curator instance,
leaving us free to manipulate Curator only for new code associated to
the new stack.

Change-Id: I0a2a08acb4c2a487663c968243fe52afedf7d4d4
Signed-off-by: Galo Navarro <galo@midokura.com>