Update patch set 3
Patch Set 3: Code-Review+2

My eyes could be deceiving me, but aren't these lines more than 100 chars?  I usually prefer not to make exceptions for something like this.  If your argument is that we should make this patch exactly the same as the one merged in MEM, then that's fine, we can merge this.  However, in that case, I would expect a new patch to address this immediately.

Patch-set: 3
Label: Code-Review=+2