Update patch set 31
Patch Set 31:

Thanks for the comments Tomohiko. So its clear the overwrite/rewrite is hard do understand. Hence, I removed it from the API, and now all ownership ops should update the ZK ownership nodes, such that they correspond to the client session that performed the last update (i.e. no more rewrite).

There is one exception: ownerUpdate()
This method has a throwIfExists argument, which - when set to true - determines the method to throw an exception if the ownership node already exists. If false, we're using the default behavior of simply updating (i.e. deleting and recreating) the ZK nodes.

So: I modified the scaladoc to reflect this new behavior:

    /**
     * Updates an object and/or its owner. The method has the following behavior
     * depending on the ownership type:
     * - for [[OwnershipType.Exclusive]], the update succeeds only if the
     *   specified owner is the current owner of the object, or if the object
     *   has been orphaned and it has no owner.
     * - for [[OwnershipType.Shared]], is is always possible to update the
     *   owner.
     * In both cases, the method rewrites the current ownership node in storage,
     * such that it corresponds to the client session performing the update.
     */
    @throws[NotFoundException]
    @throws[ReferenceConflictException]
    @throws[OwnershipConflictException]
    def update[T <: Obj](obj: T, owner: ObjId,
                         validator: UpdateValidator[T]): Unit

    /**
     * Updates the owner for the specified object. This method is similar to
     * [[update()]], except that the method throws an exception if the ownership
     * owner already exists and the throwIfExists argument is set to true.
     * The method does not modify the object data, however it increments the
     * object version to prevent concurrent modifications,
     * @param throwIfExists When set to true, the method throws an exception if
     *                      the ownership node already exists. This can be used
     *                      to prevent taking ownership from a different client
     *                      session, when the owner is the same.
     */
    @throws[NotFoundException]
    @throws[OwnershipConflictException]
    def updateOwner(clazz: Class[_], id: ObjId, owner: ObjId,
                    throwIfExists: Boolean): Unit

The methods should both succeed if the caller is the owner, unless 

As for your comment to add a method addOwner(), the updateOwner() is essentially that. I used update, because the method both adds and updates owners, but we can rename it if the name is confusing.

As for the method getWithOwnership: would this be a get that throws an exception if the provided owner is not correct?

Patch-set: 31
Label: Code-Review=0