Modify MtuIncreaser to increase MTU of UNKNOWN taps
This patch modifies the condition which is used to determine if
the MTU of the interface is increased. Previously we strictly
checked if the endpoint of interface is TUNTAP, however in
certain environments they remain UNKNOWN when the information
of the interface is notified to MtuIncreaser. To avoid is issue
and increase MTU of interfaces correctly, this patch alleviates
the condition to accept the tap interfaces with UNKNOWN endpint
if their names start with "tap".

This completely fixes MN-2784.

Change-Id: Iffe75262fd81f30b7fe1abd2c7c30e76baa784ee
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>