Add requests to NetlinkRequestBroker without writing them
This patch makes several modifications to NetlinkRequestBroker. It
makes it more cache friendly and also adds the possibility for a 3rd
type of thread to interact with it: the producer thread, which gives
work to the writer thread. This eliminates the need for an
intermediate queue between the producer thread and the
NetlinkRequestBroker.

Change-Id: Id50e387e087d50c2cfd2c77e3930bbd9d952ff9a
Signed-off-by: Duarte Nunes <duarte@midokura.com>