Update patch set 1
Patch Set 1:

(1 comment)

Thank you for your review, Ryu and Galo. I'll go with this version as adding the unique ID to port binding means that we'd need to be able to maintain such an unique ID per operation on the binding on same host, interface and port triplet. That is not the case with the Neutron side, and is cumbersome to do so on the translator side (would require a linear scan on the entire port binding data on CRUD), though it is certainly possible.

Please let me know if you think we'd better keep the original model in ZooKeeper.

Patch-set: 1