ZOOM: Support for oneof in ZOOM conversions
This patch adds support for oneof-based inheritance in ZOOM
conversions. This allows using nested Protocol Buffers messages
using the oneof keyword, to represent different classes from
the inheritance hierarchy of an object.

To allow the selection of the proper oneof fields at runtime,
the patch introduces a ZoomOneOf annotation for classes, which
specified the field name corresponding to a given class. The patch
supports nesting of any depth.

Ref: MNA-110

Change-Id: Ib0778497bca26e44ccbeee26e99e485ed56262fa
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>