Update patch set 1
Patch Set 1:

A key requirement for mm-ctl is that it exists in the hypervisor host.  It is there so that the hypervisor hosts (nova compute, docker, etc) could run the script to bind the local interface + host ID (it also requires access to the host ID properties file).  So at least in our current design, it makes sense that this is packaged as a tool of midonet agent.  'mm-ctl' is not Nova-specific.

That said, the configuration should not be mixed together with midonet agent ones.  note that the 'tasks' table, which is part of the neutron DB, IS going to be our communication channel to the cluster, so it is specific to our product.  I just think that it should have its own section in the midonet agent config (of course we could also use parameter or env variable, but a config entry is going to be very useful to have).

Patch-set: 1