Update patch set 1
Patch Set 1:

> A key requirement for mm-ctl is that it exists in the hypervisor host.  It is there so that the hypervisor hosts (nova compute, docker, etc) could run the script to bind the local interface + host ID (it also requires access to the host ID properties file). So at least in our current design, it makes sense that this is packaged as a tool of midonet agent.  'mm-ctl' is not Nova-specific.

I agree that it should be packaged with the agent, absolutely. But other than that, this tool is not interacting with the agent at all (I find the name very misleading and found several people confused about it too). The tool reads the local host ID of the agent, that's the furthest it interacts with the agent.


> That said, the configuration should not be mixed together with midonet agent ones. 

Agree.


note that the 'tasks' table, which is part of the neutron DB, IS going to be our communication channel to the cluster, so it is specific to our product.  I just think that it should have its own section in the midonet agent config (of course we could also use parameter or env variable, but a config entry is going to be very useful to have).

+1 for having the config entry. I just disagree that it should be part of the agent configuration at all. It gives the impression, misleading, that the agent needs to know about the Neutron DB. The agent today is neutron-agnostic, this makes it appear not to be.

Side note: I'm working on moving configuration to the cluster. When I'm done this discussion will be much less relevant. We can put this piece of config there (separate from the agent IMHO) so that this tool only needs to read the local ZK config and then use the config library to learn the Neutron DB address.

Patch-set: 1
Label: Verified=0