Update patch set 1
Patch Set 1:

So what you are suggesting is that we create a new configuration file for this sort of thing?  Note that this is NOT Neutron-related.  This is simply a communication channel to the cluster (it's either direct RPC or the durable queue, based on whether read or write).

Also, mm-ctl does more than just reading the host file.  It is at the core a port binding tool.  This is where the virtual ports and the physical interfaces get attached.  Whether this ought to be considered part of the agent could be debated, but it is also not part of Neutron as it should not have to deal with tap interfaces on the compute hosts (perhaps it belongs to Nova).  It just so happens that our binding mechanism goes through the cluster (and the 'tasks' table).

Consider mm-ctl as ovs-vsctl.  Instead of ovsdb, we manage zookeeper for physical to virtual port mappings.  I think making this centrally configurable is awesome.  But for this patch, let's decide on how it should be implemented.  Forget about the 'tasks' table for now, if I say, for the writes that happen from the hypervisor host to the cluster  (either mm-ctl or agent), it must go through the durable communication queue, will that section simply go into /etc/midonet/midonet.conf (we should drop 'midolman.conf') as:

[cluster]
queue_db_connection=<connection string>

Will that be a good compromise?  If you want a separate file, what will it look like?  I think it's probably better to avoid having any 'openstack' naming here since any compute virtualization will need this.

Patch-set: 1