Update patch set 1
Patch Set 1:

(1 comment)

I want to comment on the port binding logic.  Currently, our design requires that we submit the port binding information into the cluster (thus through the tasks table).  This design creates problems like this as opposed to OVS, for example, keeps the binding states locally on the host (ovsdb).  Irena, Pino or I will propose a new model in which we can eliminate having to go through the cluster to perform these bindings soon.  But for now, we need this patch so that we it is functional.  We could comment in the code that it's temporary until we design a port binding mechanism that keeps everything local to the host.

Patch-set: 1