ZOOM: Fixes internal representation of object identifiers
It replaces the internal representation of object identifiers with
strings, such that the comparison of object identifiers can be done
with object of different type (e.g. a Java UUID matching a Protocol
Buffers UUID). This ensures that ZOOM operations with Protocol
Buffer messages are passing the same CRUD unit tests as when using
ZOOM with Java classes.

The patch also corrects the call of field bindings methods, such
that they always receive an instance of the internal type of the
object's identifier.

Fixes: MNA-115

Change-Id: I860dd9d94f0c41ba3ee8ca6c860fda129615b1df
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>