Perf: Replace nmap with pktgen
This patch replaces the usage of nmap for performances tests in favour
of pktgen, which can not only send at a faster rate, but also doesn't
expect a return packets. This is important because, when using
namespaces, it is the thread executing the output packet that delivers
it to the target namespace, calculates and delivers the reply, adding
latency and a closed loop.

Change-Id: Iacd642f05fffb754fab37a79b980d8f7b1129626
Signed-off-by: Duarte Nunes <duarte@midokura.com>