Update patch set 3
Patch Set 3:

Thanks Tomohiko for comments.

I agree this is far from refined app, I wrote it on a rush for the demo when using zoom convert seemed to be the fastest way to have it do basic ops. I uploaded here in case anybody needed it for testing.

Time allowing and if we have the necessity for it, there are many improvements, using reflection to determine the available fields, maybe do autocompletion, would be welcome.

Patch-set: 3
Reviewer: Alex Bikfalvi <1002751@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0