Update patch set 21
Patch Set 21:

After looking the various Rule classes, the position is either ignored or set to 1. It seems to me that it won't be a problem to remove the position field in org.midonet.midolman.rules.Rule. However should it be removed from the api altogether? 

Regarding the Condition and its various incarnations, it seems to me also that it's fine in the sense that changing fields names has been taken care of appropriately.

Should I still create an issue to reflect changes in the other Condition and Rule classes?

Patch-set: 21