Update patch set 21
Patch Set 21:

Comments from slack chat:

I'm not sure that positions are really not used in some places. MDTS tests seem to be using position for example.

I suspect a safer option is to keep the old code as it is, and simply feed default values on the position when the data comes from a protobuf. This is simple and doesn't risk unexpected consequences if there turn out to be actual users of the position.

galo [11:55 AM]
if it turns out to require changes elsewhere

galo [11:55 AM]
whereas you could take a simple route (for now) and just keep it, and when the rule comes from the protos simply feed the position with a default value
NEW MESSAGES

Patch-set: 21