Update patch set 1
Patch Set 1:

Zookeeper config is a bit weird.  The default conf dir has a symlink to /etc/alternatives/zookeeper-conf which is a symlink back to /etc/zookeeper/conf_example.  This causes cp -a (which does NOT follow symlinks!) of the conf dirs to zookeeper.1, .2, and .3 to copy symlinks to the base config and thus get out-of-sync, with the server configuration stored in the first conf dir copied (zookeeper.1) not being updated with the server info.  We have a workaround in place (basically stop and reinit zookeeper), but this is a proper fix so we get can get rid of the workaround.

Patch-set: 1