Update patch set 7
Patch Set 7:

(11 comments)

Thanks for the comments, Tomohiko. I'm uploading the changes in the following version of the patch. I agree there is a lot of repeated and ugly code but, by now, this is the only tool we have to exercise the MM client without setting a full (and heavy) Midonet environment...

Patch-set: 7
Label: Code-Review=0