Topology API: getAll support
This patch introduces support for the getAll
operation, retrieving a snapshot with the ids
of the current topology objects of a given
type.

This patch also unifies the topology object id
type to UUID (including VTEPs).

Finally, it also unifies the protocol behavior
regarding ack/nack responses (now, all operations
except terminate send ack info - on its own or
piggy-backed on updates - to acknowledge the
reception of a request).

Ref: MNA-123

Change-Id: Id64e514bc5afea86fd289756c089ca1865237e26
Signed-off-by: Ernest Artiaga <ernest@midokura.com>