Update patch set 1
Patch Set 1:

From a quick look at the log, it looks like midolman1 is correctly setting up the flow:

KeyARP{sip=10.0.2.88, tip=10.0.2.26, op=1, sha=aa:bb:cc:00:01:01, tha=00:00:00:00:00:00}
        mask: KeyARP{sip=0.0.0.0, tip=255.255.255.255, op=255, sha=00:00:00:00:00:00, tha=00:00:00:00:00:00}
    actions: 
      SetKey{Tunnel{tun_id=10000, ipv4_src=10.0.0.8, ipv4_dst=10.0.0.128, tun_flag=0, ipv4_tos=0, ipv4_ttl=-1}}
      Output{port=3}

This looks like a flow towards the VTEP at .128, with the right VNI (10000). 

So it looks like an underlay problem, or something on the VTEP. Unfortunately there are no logs for that. (btw: it'd be useful to add the /var/log/openvswitch/*.log files to the report)

I'd verify that the VTEP container is indeed receiving traffic at the tunnel interface, and possibly check the openvswitch logs to confirm that the vtep emulator is doing what it should.

Patch-set: 1