Update patch set 1
Patch Set 1:

So far what I see there is correct. The main problem seems to be the unreachable.

According to this: 

ICMP port unreachable (22:02:08.228954 IP 10.0.0.128 > 10.0.0.8: ICMP 10.0.0.128 udp port 4789 unreachable)

This says that the container with the emulator doesn't have the management interface listening, which should be at 10.0.0.128:4789.

Can you verify if the emulator is up, and if possible send a dump of the ovsdb?

Patch-set: 1
Reviewer: Galo Navarro <1002731@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0