Change VTEP protobufs models
This change moves out all the vtep specific fields from the VxlanPort to
the VTEP and Bridge, reflecting the changes made recently for the HA
implementation.  The VxlanPort will only have a pointer to the VTEP,
where all the details (tunnel ip, tunnel zone, etc.) live.  The VNI
moves to the Network, since a VNI is 1:1 with a Network, and this is
shared in all the VxlanPort instances on a given Network.

This change just focuses on the model, so it also removes some of the
conversions from proto to the *old* simulation models, given that they
are incompatible. This is not a risk since the new storage stack is
still not in production. Later patches will focus on making the
simulation layer and vxlan gateway service compatible with the new
models.

Ref.: MNA-127
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: If439a8a12abdeadb0997848cbdef1eebe2b7e51c