Redo: Neutron's AgentMembership Translator.
Re-implements a Translator for Neutron's AgentMembership. TunnelZone
has back-references to Hosts that are its members. AgentMembership
Translator updates the back-referenes upon addition / removal of Agent
Membership in the Tunnel Zone to Hosts so that the Host's references
to its Tunnel Zones would be automatically updated by the framework.

Fix MN-3289

Change-Id: Iec5339f0f9b3fea42674b3f9d832bd13839e6768