Name gc log file in readable format
The gc log file is tagged with the number of seconds since epoch. This
makes it difficult to correlate with the run of midolman it belongs
to. This patch changes the log file to includes a human readable date,
in the format %Y%m%d_%H%M%S, e.g. gc-20150309_105352.log

Change-Id: Ie1df1402c04660744dbd19383947e655570474e8
Signed-off-by: Ivan Kelly <ivan@midokura.com>