Allocate FlowControllers structures upfront
The flow controller contains a number of large data structures, which
persist for the lifetime of the midolman instance. If GC is left to its
own devices, it will only tenure these structures once the eden region
is full, potentially leading to a very long pause if the eden region is
sufficiently large (it can be 6G+ in some installs).

This patch releases the GC pressure a bit, by allocating all the
structures when the flow controller is created, and manually triggering
a GC to tenure the data structures before starting the main midolman
loop.

The actor start timeout has also been increased.

Ref MN-3314

Change-Id: I36807a4ee5fcb0315f591f70d9bdba4c76c822d1
Signed-off-by: Ivan Kelly <ivan@midokura.com>