Missing cluster backend initialization
The MidonetBackendService was not being started
by the cluster node framework. This patch
initializes it in the cluster node and the
stand-alone services (topology api and
topology updater).

This patch also introduces the 'enabled' property
in the config files to activate the new
architecture (defaulting to false), and changes
cluster service dependencies from
MidonetBackendService to MidonetBackend (so that
it is properly instantiated as singleton by
the guice injector).

Change-Id: I2f1f5990e720e71e86dfcec137df1c595fa04166
Signed-off-by: Ernest Artiaga <ernest@midokura.com>