Update patch set 2
Patch Set 2:

The flow controller changes are actually all unnecessary. By the time we call gc() all the preallocated stuff is allocated anyhow, whether we do it in preStart or in construction.

I didn't get to the bottom of the mdts fail, but then I didn't need to, since we could get the same effect without the fc changes.

Patch-set: 2
Label: Code-Review=0