Agent host service updates interfaces
This patch adds a protocol buffers message for a host interface,
modifies the Host message to include the set of host interfaces,
and updates the host service to monitor and update the interface
list in the backend ZOOM storage.

It also enables the host service to monitor the host ownership,
and attempt retake ownership when lost, such as after a ZOOM
flush operation.

Ref: MNA-150

Change-Id: If9bfc8078e851fa3b6401503743319a2ece39f8c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>