Update patch set 15
Patch Set 15: -Code-Review

(2 comments)

I simply tried to have suggestions not to have magic numbers in the code because it slows reviews down in general. I know it's sort of inevitable and I'm not giving -1 because of that. But this would be concerns later if we have more people new to our codebase.

Patch-set: 15
Label: Code-Review=0