Implements MidoNet Task State Update.
Implements MidoNet Task State update with the C3PO. C3PO updates the
last processed MidoNet Task ID in the new 'midonet_task_state' table.
The spec is defined in the following document:

https://review.openstack.org/#/c/160229/

The 'midonet_task_state' table has only a single row with a fixed ID
as specified in the above document. C3PO periodically updates the
record with the last processed task ID recorded in the Storage.

Ref: MNA-126

Change-Id: Icf7b958a6976496c3933cdd15df9f5551b58dcfd