Implements MidoNet Data State Update.
Implements MidoNet Data State update with the C3PO. C3PO updates the
last processed MidoNet Task ID in the new 'midonet_data_state' table.
Please refer to the following documents for the details fo the specs.

https://review.openstack.org/#/c/160229/
https://review.openstack.org/#/c/164079/

This patch adopts the latest table name of 'midonet_data_state'
instead of the old name 'midonet_task_state'.  The table name change
has been proposed in the spec #164079 above, which is still being
reviewed and therefore is subject to further changes. The table has
only a single row with a fixed ID. C3PO periodically updates the
record with the last processed task ID recorded in the Storage.

Ref: MNA-126

Change-Id: Icf7b958a6976496c3933cdd15df9f5551b58dcfd